Create linkAnnotationToParam.catscript
This script creates a length parameter, adds a text annotation, and creates a formula that links the text of the annotation to the value of the parameter. The formula sets the annotation text to "Length: " followed by the value of LengthParam. When the value of LengthParam changes, the annotation text will automatically update to reflect the new value.